refactor(logic): split game_setup_create into part-of concern files (#3290)
Phase 0 prerequisite for the colonizethis_logic package split: decompose
game_setup_create.dart (637 lines, > the 500-line Phase 0 target) into
behavior-preserving `part`/`part of` concern files, mirroring the prior
Phase 0 decompositions (#3292, #3299, #3297, #3304).

- game_setup_create.dart (260): GameSetupResult, createGameFromGeneratedMaps
  orchestrator, zoom-multiplier resolution + part directives.
- game_setup_create_ownership.dart: initial OW/NW ownership assignment.
- game_setup_create_initial_game.dart: initial Game/players/market/diplomacy.
- game_setup_create_capitals.dart: capital assignment for GPs/minors/tribes.

Pure structural extraction (no behavior change); all moved members are
library-private and shared via part-of, so imports and the public API are
unchanged.

Refs #3290
